update translation processor

+++ core/processors/translations.py
@@ -28,7 +28,8 @@ class ChallengeTranslationProcessor(JobProcessor):
   async def run(self, job_id: UUID) -> TranslationProcessorResult:
     job = await self.start_job(job_id, CHALLENGE_TRANSLATION_JOB_NAME)
     try:
-      challenge, translations_to_create = await self._get_translation(job)
+      job_meta = TranslationJob(**job.meta)
+      challenge, translations_to_create = await self._get_translation(job_meta)
       if len(translations_to_create) == 0:
         await self.set_job_finished(job_id, False)
         return TranslationProcessorResult(translations=[])
@@ -38,7 +39,7 @@ class ChallengeTranslationProcessor(JobProcessor):
         input=Translation(lang="en", text=challenge.name)
       ))
 
-      result = await self._save_translations(challenge, agent_response)
+      result = await self._save_translations(job_meta, challenge, agent_response)
       await self.set_job_finished(job_id, False)
 
       return result
@@ -46,25 +47,25 @@ class ChallengeTranslationProcessor(JobProcessor):
       await self.set_job_finished(job_id, True)
       raise e
 
-  async def _get_translation(self, job: Job) -> tuple[Challenge, list[str]]:
-    job_meta = TranslationJob(**job.meta)
-
-    translations_to_create: list[str] = job_meta.langs
+  async def _get_translation(self, job_meta: TranslationJob) -> tuple[Challenge, list[str]]:
+    translations_to_create: list[str] = job_meta.langs.copy()
 
     async with self._db() as session:
       challenge_repo = ChallengeRepository(session)
-      challenge = await challenge_repo.get_challenge(job_meta.target_id, with_translations=True)
+      challenge = await challenge_repo.get_challenge(job_meta.target_id, with_translations=job_meta.append)
       if not challenge:
         raise TranslationProcessorError(f"Challenge {job_meta.target_id} not found")
 
-      for c in challenge.translations:
-        if c.lang in translations_to_create:
-          translations_to_create.remove(c.lang)
+      if job_meta.append:
+        for c in challenge.translations:
+          if c.lang in translations_to_create:
+            translations_to_create.remove(c.lang)
 
     return challenge, translations_to_create
 
   async def _save_translations(
       self,
+      job_meta: TranslationJob,
       challenge: Challenge,
       agent_response: TranslationAgentResponse
   ) -> TranslationProcessorResult:
@@ -74,16 +75,18 @@ class ChallengeTranslationProcessor(JobProcessor):
       challenge.translated_at = datetime.now(timezone.utc)
 
       challenge_repo = ChallengeRepository(session)
-      await challenge_repo.delete_old_translations(challenge.id)
+      if not job_meta.append:
+        await challenge_repo.delete_old_translations(challenge.id)
 
       for t in agent_response.translations:
         t_model = await challenge_repo.create_translation(challenge.id, t.lang, t.text)
         created_translations.append(t_model)
 
-      en_translation = next((t for t in created_translations if t.lang == "en"), None)
-      if en_translation is None:
-        t_model = await challenge_repo.create_translation(challenge.id, "en", challenge.name)
-        created_translations.append(t_model)
+      if not job_meta.append:
+        en_translation = next((t for t in created_translations if t.lang == "en"), None)
+        if en_translation is None:
+          t_model = await challenge_repo.create_translation(challenge.id, "en", challenge.name)
+          created_translations.append(t_model)
 
       await session.commit()
